+++ cloudevents/abstract/__init__.py
@@ -12,4 +12,6 @@
 #    License for the specific language governing permissions and limitations
 #    under the License.
 
-from cloudevents.abstract.event import AnyCloudEvent, CloudEvent  # noqa
+from cloudevents.abstract.event import AnyCloudEvent, CloudEvent
+
+__all__ = [AnyCloudEvent, CloudEvent]