+++ cloudevents/conversion.py
@@ -11,18 +11,6 @@
 #    WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
 #    License for the specific language governing permissions and limitations
 #    under the License.
-#
-#    Licensed under the Apache License, Version 2.0 (the "License"); you may
-#    not use this file except in compliance with the License. You may obtain
-#    a copy of the License at
-#
-#         http://www.apache.org/licenses/LICENSE-2.0
-#
-#    Unless required by applicable law or agreed to in writing, software
-#    distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
-#    WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
-#    License for the specific language governing permissions and limitations
-#    under the License.
 import datetime
 import enum
 import json
@@ -42,10 +30,10 @@ def _best_effort_serialize_to_json(
     Serializes the given value into a JSON-encoded string.
 
     Given a None value returns None as is.
-    Given a non-JSON-serializable value returns return the value as is.
+    Given a non-JSON-serializable value returns the value as is.
 
     :param value:  The value to be serialized into a JSON string.
-    :return: JSON string of the given value OR None OR given value.
+    :returns: JSON string of the given value OR None OR given value.
     """
     if value is None:
         return None