+++ cloudevents/pydantic/__init__.py
@@ -11,13 +11,7 @@
 #    WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
 #    License for the specific language governing permissions and limitations
 #    under the License.
-from cloudevents.pydantic.conversion import (  # noqa
-    from_dict,
-    from_http,
-    from_json,
-    to_binary,
-    to_dict,
-    to_json,
-    to_structured,
-)
-from cloudevents.pydantic.event import CloudEvent  # noqa
+from cloudevents.pydantic.conversion import from_dict, from_http, from_json
+from cloudevents.pydantic.event import CloudEvent
+
+__all__ = [CloudEvent, from_json, from_dict, from_http]