+++ cloudevents/tests/test_pydantic_events.py
@@ -20,7 +20,8 @@ import pytest
 from sanic import Sanic, response
 
 import cloudevents.exceptions as cloud_exceptions
-from cloudevents.pydantic import CloudEvent, from_http, to_binary, to_structured
+from cloudevents.conversion import to_binary, to_structured
+from cloudevents.pydantic import CloudEvent, from_http
 from cloudevents.sdk import converters
 from cloudevents.sdk.converters.binary import is_binary
 from cloudevents.sdk.converters.structured import is_structured
@@ -81,7 +82,6 @@ async def echo(request):
 @pytest.mark.parametrize("body", invalid_cloudevent_request_body)
 def test_missing_required_fields_structured(body):
     with pytest.raises(cloud_exceptions.MissingRequiredFields):
-
         _ = from_http(
             {"Content-Type": "application/cloudevents+json"}, json.dumps(body)
         )
@@ -186,7 +186,6 @@ def test_missing_ce_prefix_binary_event(specversion):
         "ce-specversion": specversion,
     }
     for key in headers:
-
         # breaking prefix e.g. e-id instead of ce-id
         prefixed_headers[key[1:]] = headers[key]
 
@@ -383,10 +382,10 @@ def test_cloudevent_repr(specversion):
         "ce-source": "<source-url>",
     }
     event = from_http(headers, "")
-    # Testing to make sure event is printable. I could runevent. __repr__() but
+    # Testing to make sure event is printable. I could run event. __repr__() but
     # we had issues in the past where event.__repr__() could run but
     # print(event) would fail.
-    print(event)
+    print(event)  # noqa T201
 
 
 @pytest.mark.parametrize("specversion", ["1.0", "0.3"])