+++ samples/http-json-cloudevents/client.py
@@ -16,7 +16,8 @@ import sys
 
 import requests
 
-from cloudevents.http import CloudEvent, to_binary, to_structured
+from cloudevents.conversion import to_binary, to_structured
+from cloudevents.http import CloudEvent
 
 
 def send_binary_cloud_event(url):