+++ CHANGELOG.md
@@ -5,7 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+### Added
+- A new `CloudEvent` optional `pydantic` model class is available in the 
+  `cloudevents.pydantic.event` module. The new model enables the integration of
+   CloudEvents in your existing pydantic models or integration with pydantic
+    dependent systems such as FastAPI. ([#182])
 
 ## [1.5.0] — 2022-08-06
 ### Added
@@ -180,5 +184,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#172]: https://github.com/cloudevents/sdk-python/pull/172
 [#173]: https://github.com/cloudevents/sdk-python/pull/173
 [#180]: https://github.com/cloudevents/sdk-python/pull/180
+[#182]: https://github.com/cloudevents/sdk-python/pull/182
 [#184]: https://github.com/cloudevents/sdk-python/pull/184
 [#186]: https://github.com/cloudevents/sdk-python/pull/186