+++ CHANGELOG.md
@@ -5,11 +5,23 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
 ### Added
 - A new `CloudEvent` optional `pydantic` model class is available in the 
   `cloudevents.pydantic.event` module. The new model enables the integration of
    CloudEvents in your existing pydantic models or integration with pydantic
     dependent systems such as FastAPI. ([#182])
+    
+### Changed
+- Deprecated `cloudevents.http.event_type` module, 
+    moved under `cloudevents.sdk.converters`.
+- Deprecated `cloudevents.http.json_methods` module, 
+    moved under `cloudevents.http.conversion`.
+- Deprecated `cloudevents.http.http_methods` module, 
+    moved under `cloudevents.http.conversion`.
+- Deprecated `cloudevents.http.util` module.
+ 
+
 
 ## [1.5.0] — 2022-08-06
 ### Added